修BUG

+++ Client/Assets/Script/Define/Rule.cs
@@ -511,7 +511,7 @@ public class Rule
 	// 取得傷害等級價格
 	public static int DmgLvMoney()
 	{
-		return GameDefine.iPriceDmgLv + GameDefine.iPriceDmgLv * GameDefine.fUpgradeDmgLv * PlayerData.pthis.iDamageLv;
+		return (int)(GameDefine.iPriceDmgLv + GameDefine.iPriceDmgLv * GameDefine.fUpgradeDmgLv * PlayerData.pthis.iDamageLv);
 	}
 	// 取得廣告影片贈送金額
 	public static int AdsMoney()
